Move first day of week locale call to the platform
This is platform specific, so this is where it belongs. Also this way we
only run the subprocess once, instead of every time a DateRange is used..
