+++ pytrainer/test/test_platform.py
@@ -1,0 +1,29 @@
+#Copyright (C) Nathan Jones ncjones@users.sourceforge.net
+#Copyright (C) Arto Jantunen <viiru@iki.fi>
+
+#This program is free software; you can redistribute it and/or
+#modify it under the terms of the GNU General Public License
+#as published by the Free Software Foundation; either version 2
+#of the License, or (at your option) any later version.
+
+#This program is distributed in the hope that it will be useful,
+#but WITHOUT ANY WARRANTY; without even the implied warranty of
+#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#GNU General Public License for more details.
+
+#You should have received a copy of the GNU General Public License
+#along with this program; if not, write to the Free Software
+#Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+
+import unittest
+from pytrainer.platform import get_platform
+
+class PlatformTest(unittest.TestCase):
+
+    def test_first_day_of_week_should_be_non_negative_integer(self):
+        first_day = get_platform().get_first_day_of_week()
+        self.assertTrue(first_day >= 0)
+
+    def test_first_day_of_week_should_be_less_than_seven(self):
+        first_day = get_platform().get_first_day_of_week()
+        self.assertTrue(first_day < 7)
